requirements.txt
asgiref==3.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-cors-headers==3.11.0
django-countries==7.3.2
django-filter==21.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
generics==2.0.0
gunicorn==20.1.0
idna==3.4
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.6.0
pytz==2021.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
stripe==4.1.0
typing_extensions==4.3.0
tzdata==2021.5
urllib3==1.26.12
whitenoise==6.0.0
xlwt==1.3.0
